+++ hls/test/support/file_endpoint.ex
@@ -78,7 +78,16 @@ defmodule Membrane.RTC.Engine.Support.FileSourceEndpoint do
       |> via_in(:input, toilet_capacity: 1000)
       |> child(:realtimer, Membrane.Realtimer)
       |> via_in(:input, toilet_capacity: 1000)
-      |> child(:track_sender, %StaticTrackSender{track: state.track})
+      |> child(:track_sender, %StaticTrackSender{
+        track: state.track,
+        detect_keyframe: fn buffer, track ->
+          case track.encoding do
+            :OPUS -> true
+            :H264 -> Membrane.RTP.H264.Utils.is_keyframe(buffer.payload)
+            :VP8 -> Membrane.RTP.VP8.Utils.is_keyframe(buffer.payload)
+          end
+        end
+      })
       |> bin_output(pad)
     ]