+++ hls/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.2.0-dev
 * Rename FileEndpoint to FileSourceEndpoint
+* Move StaticTrackSender to membrane_rtc_engine
 
 ## 0.1.0
 * Initial release